+++ tests/data/__init__.py
@@ -1,0 +1,8 @@
+"""Data for tests."""
+
+import importlib.resources
+
+from tests import data
+
+RADIAL_CSV_PATH = importlib.resources.files(data) / "radial.csv"
+SPIRAL_CSV_PATH = importlib.resources.files(data) / "spiral.csv"